Refactor into multiple files
Split tree implementation into multiple files. Also removed puts reference in tree_data and changed to return an array of strings that can be printed instead.
