+++ lib/trees/apple.rb
@@ -1,0 +1,11 @@
+require_relative 'fruit'
+
+class Apple < Fruit
+  attr_reader :color, :diameter
+
+  def initialize(color, diameter)
+    @color = color
+    @diameter = diameter
+  end
+end
+
