+++ lib/trees/fruit.rb
@@ -1,0 +1,8 @@
+class Fruit
+  attr_reader :has_seeds
+
+  def initialize
+    @has_seeds = true
+  end
+end
+
