+++ lib/trees/history/tree_data.rb
@@ -1,0 +1,8 @@
+class TreeData
+  attr_reader :height
+
+  def initialize(height)
+    @height = height
+  end
+end
+
